third_party/WebKit/LayoutTests/fast/dom/Document/designMode-non-html.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/editing.html#making-entire-documents-editable:-the-designmode-idl-attribute"/>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Make sure the designMode attribute is exposed on non-HTML documents.");

var xmlDocument = document.implementation.createDocument("", "", null);
shouldBe("xmlDocument.__proto__", "XMLDocument.prototype");
shouldBe("xmlDocument.__proto__.__proto__", "Document.prototype");

shouldBeEqualToString("xmlDocument.designMode", "off");
shouldNotThrow("xmlDocument.designMode = 'on'");
shouldBeEqualToString("xmlDocument.designMode", "on");
</script>
</body>
</html>
