third_party/blink/web_tests/fast/dom/Document/designMode-non-html-expected.txt

Make sure the designMode attribute is exposed on non-HTML documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xmlDocument.__proto__ is XMLDocument.prototype
PASS xmlDocument.__proto__.__proto__ is Document.prototype
PASS xmlDocument.designMode is "off"
PASS xmlDocument.designMode = 'on' did not throw exception.
PASS xmlDocument.designMode is "on"
PASS successfullyParsed is true

TEST COMPLETE
